requirements.txt
py-clob-client>=0.0.1
anthropic>=0.42.0
httpx>=0.27.0
web3>=7.0.0
discord.py>=2.4.0
numpy>=2.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
